+++ ironfish/src/network/utils/url.test.ts
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-import { parseUrl } from './parseUrl'
+import { parseUrl } from './url'
 
 describe('parseUrl', () => {
   it('should parse urls', () => {
